+++ src/plugins/XMLSpider/org/garret/perst/impl/ReplicationMasterStorageImpl.java
@@ -1,6 +1,7 @@
 package plugins.XMLSpider.org.garret.perst.impl;
 
-import plugins.XMLSpider.org.garret.perst.*;
+import plugins.XMLSpider.org.garret.perst.IFile;
+import plugins.XMLSpider.org.garret.perst.ReplicationMasterStorage;
 
 
 public class ReplicationMasterStorageImpl extends StorageImpl implements ReplicationMasterStorage
@@ -11,7 +12,7 @@ public class ReplicationMasterStorageImpl extends StorageImpl implements Replica
         this.asyncBufSize = asyncBufSize;
     }
     
-    public void open(IFile file, int pagePoolSize) {
+    public void open(IFile file, long pagePoolSize) {
         super.open(asyncBufSize != 0 
                    ? (ReplicationMasterFile)new AsyncReplicationMasterFile(this, file, asyncBufSize)
                    : new ReplicationMasterFile(this, file),
